backend/requirements.txt
fastapi
uvicorn
pandas
networkx
sqlalchemy
google-generativeai
python-dotenv
openpyxl
httpx
